JariBakken: Deprecate Element#drag_and_drop_{by,on}.
This should really have happened when RemoteWebElement was being deprecated
in the other bindings, but I wasn't aware we were removing this with no
real replacement on platforms without native events. For now, this will only
be a problem for users who call this API through the remote server.

r12575