firefox: Throw error when element cannot be clicked
Current behavior is that when element is overlapped by another one,
trying to click it will silent fail because click is happening on
overlapping element. This commit changes the behavior so instead error
is thrown telling that overlapping element would receive click. It makes
the behavior mostly similar to ChromeDriver.

Signed-off-by: Alexei Barantsev <barancev@gmail.com>